+++ packages/angular/stories/accordion.stories.ts
@@ -1,13 +1,17 @@
 import { Component } from '@angular/core';
-import { type Meta, StoryFn, moduleMetadata } from '@storybook/angular';
-import { AccordionComponent } from '../src/accordion/accordion.component';
-import { ButtonComponent } from '../src/button/button.component';
+import { AccordionComponent } from '@sl-design-system/angular/accordion';
+import { ButtonComponent } from '@sl-design-system/angular/button';
+import { type Meta, StoryObj, moduleMetadata } from '@storybook/angular';
 
 @Component({
   selector: 'sla-accordion-external-actions',
-  template: `<sl-button (click)="this.onClick('dino')">Toggle 🦖</sl-button>
+  standalone: true,
+  imports: [AccordionComponent, ButtonComponent],
+  template: ` <sl-button (click)="this.onClick('dino')">Toggle 🦖</sl-button>
     <sl-button (click)="this.onClick('alien')">Toggle 👽</sl-button>
+
     dinoState:{{ this.dinoState }} alienState:{{ this.alienState }}
+
     <sl-accordion>
       <sl-accordion-item summary="Discovering Dinosaurs: A Prehistoric Adventure" [open]="dinoState">
         Embark on a thrilling journey back in time to the age of dinosaurs! 🌎🦕🌿🦖
@@ -34,15 +38,20 @@ export class AccordionExternalActionsComponent {
 }
 
 export default {
-  title: 'Accordion',
+  title: 'Components/Accordion',
+  component: AccordionComponent,
   decorators: [
     moduleMetadata({
-      declarations: [AccordionExternalActionsComponent],
-      imports: [AccordionComponent, ButtonComponent]
+      imports: [AccordionComponent, ButtonComponent, AccordionExternalActionsComponent]
     })
   ]
 } as Meta;
 
-export const ExternalActions: StoryFn = () => ({
-  template: '<sla-accordion-external-actions></sla-accordion-external-actions>'
-});
+export const ExternalActions: StoryObj = {
+  render: () => ({
+    description: `There are scenarios where you want to be able to control the state of the accordion with an external trigger, for example when the user selects an option elsewhere on the page.
+
+              This example shows how to bind a boolean that exist in your angular component to the open state of the accordion item. It also shows how to update your boolean when the value is changed from within the accordion itself.`,
+    template: '<sla-accordion-external-actions></sla-accordion-external-actions>'
+  })
+};