+++ packages/angular/stories/routing.stories.ts
@@ -1,10 +1,9 @@
 import { NgForOf } from '@angular/common';
 import { Component } from '@angular/core';
 import { RouterLink, RouterOutlet, provideRouter, withHashLocation } from '@angular/router';
+import { MenuComponent } from '@sl-design-system/angular/menu';
+import { TabComponent, TabGroupComponent } from '@sl-design-system/angular/tabs';
 import { type Meta, StoryFn, applicationConfig, moduleMetadata } from '@storybook/angular';
-import { TabGroupComponent } from '../src/tabs/tab-group.component';
-import { TabComponent } from '../src/tabs/tab.component';
-import { MenuComponent } from '../src/menu/menu.component';
 
 @Component({
   selector: 'sla-dashboard',
@@ -113,7 +112,7 @@ export class RoutingComponent {
 }
 
 export default {
-  title: 'Routing',
+  title: 'Components/Routing',
   decorators: [
     applicationConfig({
       providers: [
@@ -155,5 +154,5 @@ export default {
 } as Meta;
 
 export const RoutingExample: StoryFn = () => ({
-  template: '<sla-routing></sla-routing>',
+  template: '<sla-routing></sla-routing>'
 });