+++ packages/components/panel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sl-design-system/panel
 
+## 0.3.0
+
+### Minor Changes
+
+- [#2842](https://github.com/sl-design-system/components/pull/2842) [`be3b2b8`](https://github.com/sl-design-system/components/commit/be3b2b8e80cb0248f445f5a865f373f6523d2457) - From now on, use `default` and `relaxed` for the `density` property.
+  The `plain` and `comfortable` values are deprecated, they will be kept for the backward compatibility for now but removed in the future.
+  Please use `default` or `relaxed` from now on.
+
+### Patch Changes
+
+- Updated dependencies [[`f676dd5`](https://github.com/sl-design-system/components/commit/f676dd52c83ef2f3429d9a393b5ec634fc05bf0e), [`d807cb2`](https://github.com/sl-design-system/components/commit/d807cb22702dc5ac1399cf0528f9ceeeb1f09f60)]:
+  - @sl-design-system/icon@1.4.1
+  - @sl-design-system/tool-bar@0.1.2
+
 ## 0.2.5
 
 ### Patch Changes