+++ packages/components/progress-bar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sl-design-system/progress-bar
 
+## 0.1.4
+
+### Patch Changes
+
+- [#2734](https://github.com/sl-design-system/components/pull/2734) [`f676dd5`](https://github.com/sl-design-system/components/commit/f676dd52c83ef2f3429d9a393b5ec634fc05bf0e) - Refactored tokens for border-radius, this was still using tokens that have now moved to the deprecated css files.
+
+- Updated dependencies [[`f676dd5`](https://github.com/sl-design-system/components/commit/f676dd52c83ef2f3429d9a393b5ec634fc05bf0e), [`d807cb2`](https://github.com/sl-design-system/components/commit/d807cb22702dc5ac1399cf0528f9ceeeb1f09f60)]:
+  - @sl-design-system/icon@1.4.1
+
 ## 0.1.3
 
 ### Patch Changes